+++ coordinator.cpp
@@ -11,7 +11,7 @@ Coordinator::Coordinator(MainGui* theGivenGui, QObject *parent) :
 
         connect(theGui,SIGNAL(newMovieNeeded(QString)),this,SLOT(LoadNewMovie(QString)));
         connect(theGui,SIGNAL(newOpencvFeedNeeded(bool)),this,SLOT(controlCameraThread(bool)));
-        connect(theGui,SIGNAL(implementNewFps(int)),this,SLOT(changeFps(int)));
+        connect(theGui,SIGNAL(implementNewFps(double)),this,SLOT(changeFps(double)));
         connect(theGui,SIGNAL(startRecording(bool,QString,QString,int)),&camBack,SLOT(StartRecording(bool,QString,QString,int)));
         connect(theGui,SIGNAL(newAvtFeedNeeded(bool)),this,SLOT(StartNewAVT(bool)));
         connect(theGui,SIGNAL(newVimbaFeedNeeded(bool)),this,SLOT(StartNewVimba(bool)));
@@ -19,7 +19,7 @@ Coordinator::Coordinator(MainGui* theGivenGui, QObject *parent) :
         connect(theGui,SIGNAL(setShutter(int)),this,SLOT(changeShutter(int)));
         connect(theGui,SIGNAL(setAutoShutter(bool)),this,SLOT(setAutoShutter(bool)));
         connect(&camBack,SIGNAL(shutterChanged(int)),theGui,SLOT(gotNewShutSpeed(int)));
-        connect(&camBack,SIGNAL(fpsChanged(int)),theGui,SLOT(gotNewFpsFromBackend(int)));
+        connect(&camBack,SIGNAL(fpsChanged(double)),theGui,SLOT(gotNewFpsFromBackend(double)));
         connect(theGui,SIGNAL(skipFrames(bool)),&camBack,SLOT(skipForwardBackward(bool)));
         connect(theGui,SIGNAL(setRoiRows(int)),&camBack,SLOT(setRoiRows(int)));
         connect(theGui,SIGNAL(setRoiCols(int)),&camBack,SLOT(setRoiCols(int)));
@@ -58,7 +58,7 @@ void Coordinator::controlCameraThread(bool startNew,QString dev)
     }
 }
 
-void Coordinator::changeFps(int newFps)
+void Coordinator::changeFps(double newFps)
 {
     if (camBack.isRunning()) {
         camBack.SetInterval(newFps);
